Fix build errors: pointer-to-int casts and missing closure_templates.h
1. Fix pointer-to-int cast errors in heap.h
   - Use uintptr_t instead of unsigned long for pointer arithmetic
   - Ensures compatibility across different architectures (32-bit/64-bit)
   - Fixes: '+nofp' feature modifier errors in boot compilation

2. Add GENHEADERS dependency to boot UEFI compilation rules
   - Ensures closure_templates.h is generated before UEFI boot code compiles
   - Fixes missing header error in /%.o rules

Fixes CI/CD build errors on both x86_64 and aarch64 platforms.

+++ src/runtime/heap/heap.h
@@ -66,8 +66,8 @@ backed_heap reserve_backed_heap_wrapper(heap meta, backed_heap parent, bytes res
 
 static inline void *page_of(void *x, bytes pagesize)
 {
-    return((void *)((unsigned long)x &
-                    (~((unsigned long)pagesize-1))));
+    return((void *)((uintptr_t)x &
+                    (~((uintptr_t)pagesize-1))));
 }
 
 static inline int subdivide(int quantum, int per, int s, int o)
